Implement ratings in organize model
You can click on the picture to select it, then move the mouse to the
top to set a rating.

You can filter the minimum rating you wish to display using the slider
on the left toolbar.

You can't yet sort by rating, nor can you view more than one day at a
time, so it's of limited usefulness at the moment, but it's an
improvement.

The UI does need work, though. If you're directly above where the
mousearea will appear when you select the picture it won't realize
you're in the rating area unless you move out and back in.