Implement dragging updates in the calendar.
I made it so that it only ever flickers when changing dates, never within
one.

It gives you 15 frames of movement after clicking in the next month before
it starts selecting the stuff under the cursor.

Its only bug is that when you drag off into the next month, then back,
it doesn't show you the first day in the current month that you entered.
It does go to that when you release it, though.