Temporary commit; rewriting parameter manager
ParameterManager had structs as the member variables of the
Q_PROPERTY's. This wasnt playing with Qt 5.5, so I'm rewriting it to
remove the ProcessingParameters structs from everything.

The ParameterManager is going to be the overall supplier of info to the
pipeline and the FilmImageProvider. There will have to be a pointer
passed in to let them call getters, and said getters will return
validity as well as the parameters themselves.

ParameterManager is also going to control validity, hopefully
simplifying the pipeline a lot.