Add filename hash appending option; other fixes
Added a SearchTable column for the time when import was started.

It used to be that the ImportTime would be that, but from now on it'll
be for when the file actually gets entered in the database.

Fixed the backup directory tooltip to actually work (it needed an
apostrophe escaped), and added a blurb to the destination directory
tooltip to say that it'll create the directory if it doesn't exist yet.

And finally for the main feature: when you check the 'Append unique
identifier' switch it'll append an underscore and 7 characters of the hash that are converted to base 62 (not a standard conversion, mind you). This is because I got two cameras of the same brand and I'm worried that the filenames will collide someday.