Implement drag and drop within the queue
It doesn't like scrolling. If you scroll and the original location goes
offscreen, it will cancel the drag. That is being really difficult to
resolve.

It's activated by a fairly long press, and the bar above the queue entry
goes to a color halfway between the "selected" white and the "active"
orange.

On hard drives, at least, the UI lags a bit when dropping an image while
it runs the database update.