Fix queue reordering bug
The delegates would move on drag and drop, but the backing data would
move too, triggering a change in the sorting that compounded with the
delegate movement.

This change separates out the actual sorted index from the
first-added incrementing index, and then copies the sorted index to the
first-added on startup.