Create import in place gui, start backend
It gives two new options: selecting individual files to import, and the
option of importing in place instead of copying to a folder structure.
They're usable in any combination.

The gui isn't quite finished, and it'll still import normally even if
you have the new options selected.

The backend is only barely started. A bit of copy paste.