Add warning for importing in place from DCIM
"DCIM" in the file/folder path, or in a subdirectory of the folder, is a
pretty good indication that you're importing from a memory card. That's
a bad thing.

This warns the user if that's happening. It highlights the text box
bright orange and provides a tooltip explaining that their photos might
be lost if they format the card.