Attempt to fix bug where queue data disappears
It's a problem when there are many many many items in the queue
(thousands). The data is dropped and scrolling in the queue yields no
new thumbnails.

I tried a bunch of fixes that didn't work.

Then I tried this and so far the bug hasn't been reproduced at all, even
though the fix itself hasn't been activated.

Strange.