Implement precise input spinner
This is something that's automatic for all continuous value sliders.

When you right click on a tool name, an iPod-like dial appears.

Drag beginning anywhere in the circle (even the inner circle)
 in order to make precise movements of the controls.

Eight full revolutions are equivalent to the full range of the slider.

Right click the tool name again to hide the dial.