Prevent saving image during computation
When switching images, clicking the save button was marking the new
image as being saved on the queue, while the old image was actually the
one being saved.

By disabling this entirely, that problem disappears.

Additionally, previously, the whole UI would freeze if the save button
was pressed during computation, until the computation finished. This
completely eliminates that problem by making it impossible to even
request saving until the computation is done.