Add date range selection to organize
In the date histogram, right-click or shift-left-click to select the
range of days from the double-click date to the one under the cursor,
including both end days.

Double-click something again to set a single day again, or use the
calendar.

The calendar doesn't show the range of days, as of yet.